+++ unity/Editor/Dialogs/UpdateDialog.cs
@@ -44,9 +44,9 @@ namespace Dissonity.Editor.Dialogs
             GUILayout.Space(10);
             GUILayout.BeginVertical();
             
-            GUILayout.Label("v1-v2.0.0 Overview", headerStyle);
+            GUILayout.Label("v2.0.1 Overview", headerStyle);
 
-            GUILayout.Label("Overhauled, improved and extended SDK.", centerStyle);
+            GUILayout.Label("Bug fixes", centerStyle);
 
             GUILayout.Space(15);
 
@@ -54,16 +54,7 @@ namespace Dissonity.Editor.Dialogs
 
             GUILayout.Space(10);
 
-            GUILayout.Label("- Dissonity.Api methods");
-            GUILayout.Label("- Dissonity.Api.Commands methods");
-            GUILayout.Label("- Dissonity.Api.Subscribe methods");
-            GUILayout.Label("- Dissonity.Api.LocalStorage methods");
-            GUILayout.Label("- Dissonity.Api.Proxy methods");
-            GUILayout.Label("- Dissonity.Api.HiRpc methods");
-            GUILayout.Label("- Dissonity.Utils methods");
-            GUILayout.Label("- Dissonity.Models classes");
-            GUILayout.Label("- @DiscordMock object");
-            GUILayout.Label("- @JavascriptMock object");
+            GUILayout.Label("- Added link.xml file to the Dissonity folder.");
 
             GUILayout.Space(15);
 
@@ -71,9 +62,7 @@ namespace Dissonity.Editor.Dialogs
 
             GUILayout.Space(10);
 
-            GUILayout.Label("- Playtest your Discord activity within Unity!");
-            GUILayout.Label("- Not limited to Node.js servers.");
-            GUILayout.Label("- You must call and await Api.Initialize once per runtime.");
+            GUILayout.Label("- Api.Commands.ShareLink returns a complete mock response.");
 
             GUILayout.Space(15);
 
@@ -81,14 +70,13 @@ namespace Dissonity.Editor.Dialogs
 
             GUILayout.Space(10);
 
-            GUILayout.Label("- Fixed check to detect absolute urls in proxy requests.");
-            GUILayout.Label("- Added hiRPC mock runtime limitations.");
+            GUILayout.Label("- Api.GuildId is now nullable as it should be.");
 
             GUILayout.Space(10);
 
-            GUILayout.Label("And more!", italicStyle);
+            //GUILayout.Label("And more!", italicStyle);
 
-            GUILayout.Space(20);
+            //GUILayout.Space(20);
 
             if (GUILayout.Button("Got it!", GUILayout.Height(25)))
             {
@@ -98,17 +86,6 @@ namespace Dissonity.Editor.Dialogs
             GUILayout.Space(15);
 
             GUILayout.Label("Links", subHeaderStyle);
-
-            //todo remove after initial release
-            if (EditorGUILayout.LinkButton("Full Alpha changelog"))
-            {
-                Application.OpenURL("https://github.com/Furnyr/Dissonity/blob/v2/.github/archive/ALPHA_CHANGELOG.md");
-            }
-
-            if (EditorGUILayout.LinkButton("Full Beta changelog"))
-            {
-                Application.OpenURL("https://github.com/Furnyr/Dissonity/blob/v2/.github/archive/BETA_CHANGELOG.md");
-            }
 
             if (EditorGUILayout.LinkButton("Documentation"))
             {