+++ unity/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this package will be documented in this file.
 
+## [2.0.1] - 2025-05-30
+
+### Added
+
+- Added link.xml to the Dissonity folder. This fixes stripping issues in some Unity versions.
+
+### Changed
+
+- Updated the mock return data of `Api.Commands.ShareLink`.
+
+### Fixed
+
+- `Api.GuildId` is now nullable, since it has a null value in DMs.
+
 ## [2.0.0] - 2025-04-15
 
 Compared to `v2.0.0-beta.2`: